requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
Click==7.0
cryptography==2.7
Flask==1.1.1
Flask-SQLAlchemy==2.1
flask-swagger-ui==3.20.9
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2==2.8.3
pycparser==2.19
PyJWT==1.7.1
PyYAML==5.1.2
six==1.12.0
SQLAlchemy==1.3.6
Werkzeug==0.15.5
